% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostFunction.R
\name{cout}
\alias{cout}
\title{CostFunction}
\usage{
cout(x, y, theta)
}
\arguments{
\item{x}{Matrix of explanatory variables to predict}

\item{y}{Target Variable Matrix}

\item{theta}{Coefficient matrix for logistic regression}
}
\value{

}
\description{
A loss function is a method of evaluating how well your algorithm models your datasets
If they're pretty good, it will output a lower number
}
\examples{
cout(x,y,coef)
}
