% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_online_batch.R, R/fit_online_batch_Final.R
\name{fit}
\alias{fit}
\title{Fit Function}
\usage{
fit(
  formula,
  data,
  mode,
  batch_size,
  ncores = 0,
  coef,
  max_iter = 100,
  learningrate = 0.1,
  tol = 1e-04
)

}
\arguments{
\item{formula}{Target variable ~ explanatory variables}

\item{data}{dataframe required to perform the gradient descent}

\item{mode}{Gradient descent mode (Online, Mini-Batch, Batch)}

\item{batch_size}{size of the cutting to be carried out for the "mini-batch" mode}

\item{ncores}{Number of cores required for parallelization}

\item{coef}{Fixed number for creating the base vector}

\item{max_iter}{Max number of iterations not to be exceeded}

\item{learningrate}{Learning rate which determines the step of the descent}

\item{tol}{Tolerance threshold for which convergence is said to be "finished"}
}
\value{



}
\description{
Implementation of binary logistic regression with stochastic gradient descent for algorithm

}
