% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_xy.R
\name{gradient_xy}
\alias{gradient_xy}
\title{Gradient Function}
\usage{
gradient_xy(data, theta)
}
\arguments{
\item{data}{dataframe required to perform the gradient descent}

\item{theta}{Coefficient matrix for logistic regression}
}
\value{
Returns a gradient vector
}
\description{
The gradient of a function f for the parallelisation, denoted as ∇f, is the collection of all its partial derivatives into a vector
}
\examples{
gradient_xy(data_ex, coef_ex)
}
