% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(objet, newdata, type)
}
\arguments{
\item{objet}{Matrix of explanatory variables to predict}

\item{newdata}{Target Variable Matrix}

\item{type}{Coefficient matrix for logistic regression}
}
\value{

}
\description{
A loss function is a method of evaluating how well your algorithm models your datasets
If they're pretty good, it will output a lower number
}
\examples{
cout(x,y,coef)
}
