% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{matrix of explanatory variables to predict}
}
\value{
Returns a vector of positive / negative class membership probabilities
}
\description{
The sigmoid function is also called a squashing function as its domain is the set of all real numbers, and its range is (0, 1)
}
\examples{
sigmoid(x)
}
