% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(objet, newdata, type)
}
\arguments{
\item{objet}{the S3 object that the fit function returns}

\item{newdata}{dataframe explanatory variables on which to apply the prediction}

\item{type}{Output type "class" (For predicted classes) or "posterior" (for the probability of class memberships)}
}
\value{

}
\description{
prediction Binary or probabilities prediction
}



